requirements.txt
Appium-Python-Client==5.2.4
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
et_xmlfile==2.0.0
h11==0.16.0
idna==3.11
iniconfig==2.1.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.6
MarkupSafe==3.0.3
ml_dtypes==0.5.0
MouseInfo==0.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.2
pytest-html==4.1.1
pytest-metadata==3.1.1
pytweening==1.2.0
requests==2.32.5
rubicon-objc==0.5.2
selenium==4.37.0
sentencepiece==0.2.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.2.0
